backend/pyproject.toml
[tool.poetry]
name = "bim-ai-backend"
version = "0.1.0"
description = "Backend for BIM-AI Multi-Agent System"
authors = ["BIM-AI Team"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
pydantic = "^2.6.0"
pymupdf = "^1.23.0"
google-cloud-aiplatform = "^1.38.0"
google-generativeai = "^0.3.0"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.9"
aiofiles = "^23.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
